smart_complaint_register/requirements.txt
streamlit==1.28.1
google-generativeai==0.3.0
pymongo==4.6.0
python-dotenv==1.0.0
pandas==2.1.4
langchain==0.0.354
langchain-google-genai==0.0.5
bson==0.5.10
dnspython==2.4.2